+++ packages/data/src/index.ts
@@ -9,6 +9,8 @@ export * from "./loaders";
 export * from "./dc-metro-geojson";
 export * from "./us-metros-geojson";
 export * from "./us-states";
+export * from "./us-counties";
+export * from "./zip-to-county";
 export * from "./metro-shards";
 export * from "./metro-tiles";
 export * from "./locale-quotes";